Remove deprecated DateUtils class
